+++ src/main/java/org/park/dtos/payments/PaymentResponseDTO.java
@@ -3,11 +3,12 @@ package org.park.dtos.payments;
 import org.park.model.enums.PaymentMethod;
 import org.park.model.enums.PaymentStatus;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PaymentResponseDTO(
         UUID id,
-        double totalAmount,
+        BigDecimal totalAmount,
         PaymentMethod paymentMethod,
         PaymentStatus paymentStatus
 ) {
